% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt.R
\name{filter_and_estimate}
\alias{filter_and_estimate}
\title{Utility function for filtering and estimating}
\usage{
filter_and_estimate(incidence, n_resample, level = 0.95, shift_amt = 0, n = 7)
}
\arguments{
\item{incidence}{incidence timeseries, vector}

\item{n_resample}{number of bootstrapping attempts for confidence interval estimation}

\item{level}{confidence interval level}

\item{shift_amt}{number of timesteps to shift estimation back by (integer)}

\item{n}{number of timesteps to filter (longer is smoother), must be odd}
}
\value{
dataframe with estimated r(t)
}
\description{
Utility function for filtering and estimating
}
