% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{linear_filter}
\alias{linear_filter}
\title{Linear Savitzky-Golay filter with confidence intervals}
\usage{
linear_filter(incidence, N = 7, level = 0.95)
}
\arguments{
\item{incidence}{Incidence timeseries, vector}

\item{N}{number of timesteps to filter (longer is smoother), must be odd}

\item{level}{confidence interval level}
}
\value{
dataframe with filtered timeseries and confidence intervals
}
\description{
Filter an timeseries using a Savitzky-Golay filter
}
